server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
email-validator==2.1.0
PyJWT==2.8.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9  # For PostgreSQL support (optional)
itsdangerous==2.1.2
Jinja2==3.1.2
click==8.1.7
